Somewhat working CPU
CPU executes basic VGA fill routine (QTest). Meaning of JMP had to be changed, because it is used in RTS, and return address from stack cannot be added to the immediate value P is pointing to. In other words, BRANCH must be used for JUMP, and JUMP can only be used when the X, Y, S is used for target address calculation.